Change to a custom package
Resolves #3
