--- Assets/Scripts/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fc3f763c5ec705846b197e7603f66222
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
