+++ Editor.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 2cda990e2423bbf4892e6590ba056729
+guid: fcee666c56f0f0443b2b11ce39d93a69
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
