+++ Editor/SettingsScriptableObject.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 83f028dc9a2e0974abf1dd832d740c4a
+guid: 28f252b864b623a47b442aa1153256f6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
