+++ Editor/Unity.Build-Pipeline.Editor.asmdef
@@ -1,0 +1,15 @@
+{
+    "name": "Unity.Build-Pipeline.Editor",
+    "references": [],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
