+++ Editor/Unity.Build-Pipeline.Editor.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 8e68125f5c7094c41be13eab7c456f45
-folderAsset: yes
-DefaultImporter:
+guid: 670e0d55767746841b804075af6441fb
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
