--- ProjectSettings/AudioManager.asset
@@ -1,19 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!11 &1
-AudioManager:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Volume: 1
-  Rolloff Scale: 1
-  Doppler Factor: 1
-  Default Speaker Mode: 2
-  m_SampleRate: 0
-  m_DSPBufferSize: 1024
-  m_VirtualVoiceCount: 512
-  m_RealVoiceCount: 32
-  m_SpatializerPlugin: 
-  m_AmbisonicDecoderPlugin: 
-  m_DisableAudio: 0
-  m_VirtualizeEffects: 1
-  m_RequestedDSPBufferSize: 1024
