--- ProjectSettings/NetworkManager.asset
@@ -1,8 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!149 &1
-NetworkManager:
-  m_ObjectHideFlags: 0
-  m_DebugLevel: 0
-  m_Sendrate: 15
-  m_AssetToPrefab: {}
