--- ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,0 @@
-m_EditorVersion: 2019.3.11f1
-m_EditorVersionWithRevision: 2019.3.11f1 (ceef2d848e70)
