--- ProjectSettings/TimeManager.asset
@@ -1,9 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!5 &1
-TimeManager:
-  m_ObjectHideFlags: 0
-  Fixed Timestep: 0.02
-  Maximum Allowed Timestep: 0.1
-  m_TimeScale: 1
-  Maximum Particle Timestep: 0.03
