--- ProjectSettings/UnityConnectSettings.asset
@@ -1,34 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!310 &1
-UnityConnectSettings:
-  m_ObjectHideFlags: 0
-  serializedVersion: 1
-  m_Enabled: 0
-  m_TestMode: 0
-  m_EventOldUrl: https://api.uca.cloud.unity3d.com/v1/events
-  m_EventUrl: https://cdp.cloud.unity3d.com/v1/events
-  m_ConfigUrl: https://config.uca.cloud.unity3d.com
-  m_TestInitMode: 0
-  CrashReportingSettings:
-    m_EventUrl: https://perf-events.cloud.unity3d.com
-    m_Enabled: 0
-    m_LogBufferSize: 10
-    m_CaptureEditorExceptions: 1
-  UnityPurchasingSettings:
-    m_Enabled: 0
-    m_TestMode: 0
-  UnityAnalyticsSettings:
-    m_Enabled: 0
-    m_TestMode: 0
-    m_InitializeOnStartup: 1
-  UnityAdsSettings:
-    m_Enabled: 0
-    m_InitializeOnStartup: 1
-    m_TestMode: 0
-    m_IosGameId: 
-    m_AndroidGameId: 
-    m_GameIds: {}
-    m_GameId: 
-  PerformanceReportingSettings:
-    m_Enabled: 0
