--- ProjectSettings/VFXManager.asset
@@ -1,12 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!937362698 &1
-VFXManager:
-  m_ObjectHideFlags: 0
-  m_IndirectShader: {fileID: 0}
-  m_CopyBufferShader: {fileID: 0}
-  m_SortShader: {fileID: 0}
-  m_StripUpdateShader: {fileID: 0}
-  m_RenderPipeSettingsPath: 
-  m_FixedTimeStep: 0.016666668
-  m_MaxDeltaTime: 0.05
