+++ package.json.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: c3ca06dfc3df78442a15094b5f0f1290
-folderAsset: yes
-DefaultImporter:
+guid: 97f6ecac9af56b44abff1d68ce8d6d58
+PackageManifestImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
